// src/app/business-logic/model/workers/workersGetActivityReportRequest.ts
/**
 * workers
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.14
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface WorkersGetActivityReportRequest {
    /**
     * Starting time (in seconds from epoch) for collecting statistics
     */
    from_date: number;
    /**
     * Ending time (in seconds from epoch) for collecting statistics
     */
    to_date: number;
    /**
     * Time interval in seconds for a single statistics point. The minimal value is 1
     */
    interval: number;
}